fix: bug with won state and auto updating of the URLs
